Have Rust handle sorting the process list
We need to sort the process list by the RSS column to determine which
task was using the most memory when the oom-killer triggered. We already
recreate the "obvious" way someone might do this by passing the output
to the 'sort' command. While the Command module does let us call out to
the OS (and even pipe commands) we can do better.

The ps list from the oom kill is already in our program's memory. With a
little effort, we can convert the list to a two-dimensional matrix of
strings and then sort them by position.