Report a running total of memory used by commands
Build a map that tracks the running total of RSS in use by unique
command names. A BTreeMap will keep the keys sorted for us, but
that's not quite enough for our purposes. The map needs to be
converted back to a vector in order to sort the values.

With this change in place we no longer have to outsource the final
reporting to another operating system utility.