Add flexibility to match end of oom-kill report
Not all oom-kill reports end with the string

    "Out of memory: Kill process XXX ... or sacrifice child"

We also need to consider other variations, such as a memory cgroup
triggering the oom-kill. Ideally we could just ignore all of the
possible variations, but we need to know where the oom-kill report ends
so we know where to stop our match.

So adjust the regex to minimally match the string "out of memory" as
the (case-insensitive) end of report barrier and change the capturing
group for the full report we want to process. (Note that we can still
end up capturing more of the log than we want, which can corrupt the
process list. We'll check for that when we adjust parsing the ps
list.)