Remove summary lines from end of report
There are several different ways an oom-kill report may end, depending
on what triggered the oom-killer. (These are the string entries after
the process list.)

Look for lines that may still have end of oom-kill "noise" in them
before proceeding to parse the process list. With these lines dropped
we don't have to do duplicate work looking for the end of report
before sorting the PIDs.