Display just the top 20 processes consuming RAM
First, separate the section that sorts and prints individual commands
from the code for sorting and printing the ps list.

The vector 'v' already has the sorted process list. Because we were
already iterating over each entry and collecting it into a new string 's'
we couldn't also take the top 20 entries from the vector 'v'. But now,
instead of building a new string with the results, we have a lot more
flexibility both for how we use the vector 'v' and for formatting the
end result.