Print process list directly from oom kill
The process list in an oom kill report changes between kernel releases.
Fortunately, the PID, RSS, and process name columns are always present
and in the same column positions, so no logic is effected. But we have
to do some extra work to capture the header from the report instead of
just printing a generic one.

Fortunately most of the work is already done. We can just change our
groupings in the regex already in place to capture the process list.