Compile end of oom-killer regex only once
The pattern to match the end of the oom killer report was being built
every time through the loop. This is a silly, but not entirely
meaningless optimization for a release build.

Old version: Calling Regex:new() each time through the loop.

    % cargo build --release
    Compiling parseoom v0.1.0 (/home/ewk/Projects/parseoom)
        Finished release [optimized] target(s) in 1.06s

    % time target/release/parseoom messages
    Total RAM: 2047.5 MiB
    Free swap  = 0kB

    --->8

    real    0.02s
    user    0.02s
    sys     0.00s

New version: Calling Regex:new() once before entering the loop.

    % cargo build --release
    Compiling parseoom v0.1.0 (/home/ewk/Projects/parseoom)
        Finished release [optimized] target(s) in 1.06s

    % time target/release/parseoom messages
    Total RAM: 2047.5 MiB
    Free swap  = 0kB

    --->8

    real    0.01s
    user    0.01s
    sys     0.00s