Skip duplicate end of report delimiters
Variations of the string "Out of memory: " mark the end of the oom-kill
report. In some situations, like memory pressure, the logging may print
this line multiple times or print the end of another report. We want,
and always intended, to stop reading data at the first occurrence of
this string.

To be *sure* we stop on the first occurrence, the biggest change is just
to use a non-greedy operator to grab the text in between the two strings
designated as the report delimiters.

  - Turn off string matches newline before matching "[oO]ut of memory:"
  - Use a non-greedy operator to match leading text on that line.
  - Group the "[oO]ut of memory:" string and explicitly match it once.